Keep existing entries in /etc/resolv.conf
Instead of rewriting /etc/resolv.conf with just one nameserver entry,
pointing to our private named, merely add our private named to the top
of resolv.conf, so it will be the first resolver queried.

If our private named server is not functional, at least the fallbacks
will be queried so public hostnames can resolve.  This is especially
helpful at installation-time when the named host may not yet be
configured but we want other hosts to still be able to download their
needed packages.