Add configuration for JBoss entitlements
Add configure_jbosseap_subscription and configure_jbossews_subscription
functions, which must be modified to enable the corresponding channels
if one wants to be able to install the corresponding cartridges.

Modify the converter script to modify those functions to use internal
servers.

Rename configure_jboss_cartridge_repo to
configure_jbosseap_cartridge_repo for consistency.  (The JBossEWS
cartridge is in the Node channel.)