Refactor code into parse_cmdline and set_defaults
parse_cmdline is a new function that now contains the code to parse
/proc/cmdline and set the CONF_* variables.

set_defaults is a new function that now contains the code that sets the
various variables used in the script, based on the values of the CONF_*
variables and reasonable defaults.

This change should add a little more structure to the kickstart script
and make it easier to disable the cmdline parsing when the %post section
is extracted for use as a shell script.