Add configure_mongo_auth_plugin
Rename configure_auth_plugin to configure_remote_user_auth_plugin.

Add configure_mongo_auth_plugin, which configures the broker to use the
old MongoDB-based authentication plug-in and adds the admin user to the
database.

Add a commented-out invocation of configure_mongo_auth_plugin to make it
easy to switch from the remote-user plug-in to the MongoDB-based plug-in
for testing.