openshift-amz: Only grab DNSSEC key if needed
Only use ssh to get the DNSSEC key from the host running named if that
host is a remote host.  If we are installing the broker on the same host
on which we installed named, then we already have the key still from
configuring named.

With this commit, DNS should work automatically in new devenv instances.