Add and use update-ip-address.sh for devenv launch
Add the script update-ip-address.sh in the misc/ folder.  This script
checks the value of PUBLIC_IP in /etc/openshift/node.conf to get the IP
address with which the host has been configured.  It then checks the
current IP address of the host, and if that address is not the same, it
updates several configuration files with the appropriate IP address and
restarts services.

This script is necessary for devenv launch because launch creates an
ami instance that has a different IP address than the one with which it
was configured at build time.

Note that we need to rebuild the ami to include the update-ip-address.sh
script before it will be available to launch.