devenv: enable tests
Copy over the test code from the Origin devenv script.  Many tests must
remain disabled for the time being: Some tests rely on a fix to support
the remote-auth plug-in that needs to be merged in enterprise-server,
and Cucumber tests also require a fix before any of them will work.