Don't get ruby193-rubygems-devel from devenv repo
Prior to this commit, we get ruby193-rubygems from the Enterprise repos
and ruby193-rubygems-devel from the devenv repo.  In doing so, we get
incompatible versions of the two packages.  After this commit, we should
be getting both packages from the Enterprise repos so that both packages
can be installed successfully.