<1.2 upgrade> copy name from old gear to fix bug 971355 and bug 971354
https://bugzilla.redhat.com/show_bug.cgi?id=971354
https://bugzilla.redhat.com/show_bug.cgi?id=971355
migrate-mongo-2.0.23 = model refactor migration

During the model refactor migration, a new gear is created from the old
one. But the name of the gear isn't copied over explicitly, so the
default is used. For secondary gears, at the time the codebase default
was to use the first ten chars of the gear id. Later the codebase
switched the default to the whole gear id. Explicitly copying the gear
name should resolve this.