<upgrade> Bug 988478, fix up mcollective conf to use activemq after upgrade
This change adds logic to convert the existing mcollective server
config on nodes and client config on brokers to use activemq instead
of stomp as the connector plugin. It handles both the single-server
case and the server pool case.

Hardcoded paths to the mcollective client and server config files have
been replaced with variable substitution