Adding the oo-admin-yum-validator manpage from upstream
Generated with:

txt2man -t oo-admin-yum-validator -s 8 oo-admin-yum-validator.txt2man > oo-admin-yum-validator.8

We're forced to update the version number whenever we make changes because of
the way tito generates the source tarball.  The Errata Tool will block the
release if a source tarball changes but the version number doesn't.  I put an
'a' at the end to allow is to still have a mostly meaningful version number and
allow us to continue iterating.