<tito props> use ReleaseTagger and UpstreamBuilder
For openshift-enterprise-release and related RPMs, we want to bump the
release very deliberately, not incidentally. So, by default, to build
something with tito, this will bump the release and build a patch
against the base release tarball.

There are some big drawbacks to this due to what patching can and cannot
handle on RHEL 6. In any of the following situations, it will be
necessary to bump the version manually and tito tag --keep-version:
* We actually want to change the version
* There is no base release (x.y.z-1)
* The patch involves anything that isn't a simple text change
  * deleting files or dirs
  * symlinks
  * changing a binary file
In most cases, tito build/release will fail, but in at least one case,
running tito release on a Fedora host worked but the patch silently
failed in the RHEL 6 brew build. So be careful.